+++ src/pages/search/SearchDrawer/LocationSelection.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useAppDispatch } from '@store';
 import { getWeatherByGeolocation } from '@slices/weatherSlice';
-import { LocationResults } from '@pages/Main/SearchDrawer/LocationResults';
+import { LocationResults } from '@pages/search/SearchDrawer/LocationResults';
 
 import { ButtonWithIcon } from '@components/ButtonWithIcon';
